+++ app/src/withCall/java/com/example/chat/application/chatexample/CallContract.java
@@ -2,6 +2,8 @@ package com.example.chat.application.chatexample;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.support.annotation.RequiresApi;
+import android.support.annotation.RestrictTo;
 
 import com.fanap.podcall.view.CallPartnerView;
 import com.fanap.podchat.call.contacts.ContactsFragment;
@@ -41,7 +43,6 @@ public interface CallContract {
         }
 
 
-
         default void onAddParticipant() {
         }
 
@@ -52,8 +53,6 @@ public interface CallContract {
         }
 
 
-
-
         default void onState(String state) {
         }
 
@@ -62,108 +61,175 @@ public interface CallContract {
         }
 
 
-        default void onTokenExpired(){}
+        default void onTokenExpired() {
+        }
 
-        default void showCallRequest(String callerName){}
+        default void showCallRequest(String callerName) {
+        }
 
-        default void onVoiceCallRequestRejected(String callerName){}
+        default void onVoiceCallRequestRejected(String callerName) {
+        }
 
-        default void onVoiceCallEnded(String uniqueId, long subjectId){}
+        default void onVoiceCallEnded(String uniqueId, long subjectId) {
+        }
 
-        default void onVoiceCallStarted(String uniqueId, String clientId){}
+        default void onVoiceCallStarted(String uniqueId, String clientId) {
+        }
 
-        default void onGetCallHistory(List<CallVO> calls){}
+        default void onGetCallHistory(List<CallVO> calls) {
+        }
 
-        default void onCallReconnect(long callId){}
+        default void onCallReconnect(long callId) {
+        }
 
-        default void onCallConnect(long callId){}
+        default void onCallConnect(long callId) {
+        }
 
-        default void onCallDelivered(CallDeliverResult result){}
+        default void onCallDelivered(CallDeliverResult result) {
+        }
 
-        default void onGroupVoiceCallRequestReceived(String callerName, String title, List<Participant> participants){}
+        default void onGroupVoiceCallRequestReceived(String callerName, String title, List<Participant> participants) {
+        }
 
-        default void onCallParticipantLeft(String response){}
+        default void onCallParticipantLeft(String response) {
+        }
 
-        default void onCallParticipantJoined(String response){}
+        default void onCallParticipantJoined(String response) {
+        }
 
-        default void onCallParticipantRemoved(String name){}
+        default void onCallParticipantRemoved(String name) {
+        }
 
-        default void onRemovedFromCall(){}
+        default void onRemovedFromCall() {
+        }
 
-        default void updateStatus(String message){}
+        default void updateStatus(String message) {
+        }
 
-        default void onThreadClosed(long subjectId){}
+        default void onThreadClosed(long subjectId) {
+        }
 
-        default void onCallCreated(long threadId){}
+        default void onCallCreated(long threadId) {
+        }
 
-        default void audioCallMuted(){}
-        default void audioCallUnMuted(){}
+        default void audioCallMuted() {
+        }
 
-        default void callParticipantMuted(CallParticipantVO participant, CallPartnerView partnerView){}
-        default void callParticipantUnMuted(CallParticipantVO participant,CallPartnerView partnerView){}
+        default void audioCallUnMuted() {
+        }
 
-        default void audioCallMutedByAdmin(){}
-        default void audioCallUnMutedByAdmin(){}
+        default void callParticipantMuted(CallParticipantVO participant, CallPartnerView partnerView) {
+        }
 
+        default void callParticipantUnMuted(CallParticipantVO participant, CallPartnerView partnerView) {
+        }
 
-        default void callParticipantCanceledCall(String name){}
+        default void audioCallMutedByAdmin() {
+        }
 
-        default void hideCallRequest(){}
-
-        default void showContactsFragment(ContactsFragment contactsWrappers){}
-
-        default void updateContactsFragment(ArrayList<ContactsWrapper> contactsWrappers){}
-
-        default void onGetSentryLogs(String logs){}
-
-        default void onChatProfileUpdated(ResultUpdateProfile result){}
-
-        default void onLoginNeeded(){}
-
-        default void hideFabContactButton(){}
-
-        default void setInitState(){}
-
-        default void onScreenIsSharing(){}
-
-        default void onCallParticipantSharedScreen(){}
-
-        default void onScreenShareEnded(){}
-
-        default void onCallParticipantStoppedScreenSharing(){}
-
-        default void onCallRecordingStarted(){}
-        default void onCallRecordingStopped(){}
-        default void onParticipantStartedRecordingCall(String name){}
-        default void onParticipantStoppedRecordingCall(String name){}
-
-        default void hideContactsFragment(){}
-        default void showFabContact(){}
-
-        default void updateTvCallee(String txt){}
-
-        default void updateTvCaller(String txt){}
-
-        default void showMessage(String msg){}
-
-        default void switchToRecentCallsLoading(){}
-
-        default void hideCameraPreview(){}
-        default void showCameraPreview(){}
+        default void audioCallUnMutedByAdmin() {
+        }
 
 
-        default void hideRemoteViews(){}
-        default void showRemoteViews(){}
+        default void callParticipantCanceledCall(String name) {
+        }
 
-        default void showVideoCallElements(){}
+        default void hideCallRequest() {
+        }
 
-        default void updateCallerImage(String profileUrl){}
-        default void updateCallImage(String profileUrl){}
+        default void showContactsFragment(ContactsFragment contactsWrappers) {
+        }
+
+        default void updateContactsFragment(ArrayList<ContactsWrapper> contactsWrappers) {
+        }
+
+        default void onGetSentryLogs(String logs) {
+        }
+
+        default void onChatProfileUpdated(ResultUpdateProfile result) {
+        }
+
+        default void onLoginNeeded() {
+        }
+
+        default void hideFabContactButton() {
+        }
+
+        default void setInitState() {
+        }
+
+        default void onScreenIsSharing() {
+        }
+
+        default void onCallParticipantSharedScreen() {
+        }
+
+        default void onScreenShareEnded() {
+        }
+
+        default void onCallParticipantStoppedScreenSharing() {
+        }
+
+        default void onCallRecordingStarted() {
+        }
+
+        default void onCallRecordingStopped() {
+        }
+
+        default void onParticipantStartedRecordingCall(String name) {
+        }
+
+        default void onParticipantStoppedRecordingCall(String name) {
+        }
+
+        default void hideContactsFragment() {
+        }
+
+        default void showFabContact() {
+        }
+
+        default void updateTvCallee(String txt) {
+        }
+
+        default void updateTvCaller(String txt) {
+        }
+
+        default void showMessage(String msg) {
+        }
+
+        default void switchToRecentCallsLoading() {
+        }
+
+        default void hideCameraPreview() {
+        }
+
+        default void showCameraPreview() {
+        }
+
+
+        default void hideRemoteViews() {
+        }
+
+        default void showRemoteViews() {
+        }
+
+        default void showVideoCallElements() {
+        }
+
+        default void updateCallerImage(String profileUrl) {
+        }
+
+        default void updateCallImage(String profileUrl) {
+        }
+
+
+        default void addNewView(CallPartnerView partnerView) {
+        }
     }
 
     interface presenter {
 
-        void enableAutoRefresh(Activity activity,String entry);
+        void enableAutoRefresh(Activity activity, String entry);
 
         void setToken(String token);
 
@@ -247,5 +313,8 @@ public interface CallContract {
         void turnOnCamera();
 
         void turnOffCamera();
+
+        @RestrictTo(RestrictTo.Scope.TESTS)
+        void prepareNewView(CallPartnerView partnerView);
     }
 }