+++ app/src/withCall/java/com/example/chat/application/chatexample/CallPresenter.java
@@ -4,6 +4,7 @@ package com.example.chat.application.chatexample;
 import static com.example.chat.application.chatexample.ServerConfig.*;
 
 
+import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.Application;
 import android.app.NotificationManager;
@@ -12,6 +13,7 @@ import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
+import android.support.annotation.Nullable;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
@@ -59,9 +61,11 @@ import com.fanap.podchat.call.result_model.JoinCallParticipantResult;
 import com.fanap.podchat.call.result_model.LeaveCallResult;
 import com.fanap.podchat.call.result_model.MuteUnMuteCallParticipantResult;
 import com.fanap.podchat.call.result_model.RemoveFromCallResult;
+import com.fanap.podchat.chat.CallPartnerViewManager;
 import com.fanap.podchat.chat.Chat;
 import com.fanap.podchat.chat.ChatAdapter;
 import com.fanap.podchat.chat.ChatHandler;
+import com.fanap.podchat.chat.MainThreadExecutor;
 import com.fanap.podchat.chat.ping.request.StatusPingRequest;
 import com.fanap.podchat.chat.thread.respone.CloseThreadResult;
 import com.fanap.podchat.chat.user.profile.ResultUpdateProfile;
@@ -100,10 +104,11 @@ import com.google.gson.GsonBuilder;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 
-public class CallPresenter extends ChatAdapter implements CallContract.presenter, Application.ActivityLifecycleCallbacks {
+public class CallPresenter extends ChatAdapter implements CallContract.presenter, Application.ActivityLifecycleCallbacks, CallPartnerViewManager.IAutoGenerate {
 
     public static final int CALL_PERMISSION_REQUEST_CODE = 101;
 
@@ -135,6 +140,8 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     private List<CallPartnerView> remotePartnersViews;
     private CallPartnerView cameraPreview;
     private int cameraId = CameraId.FRONT;
+    @Nullable
+    CallPartnerViewManager cpvManager;
 
 
     public CallPresenter(Context context, CallContract.view view, Activity activity) {
@@ -253,10 +260,10 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
                         .build();
 
         AudioCallParam audioCallParam = new AudioCallParam.Builder()
-                .setFrameRate(16000) //8000
-                .setBitrate(12000) //8000
+                .setFrameRate(8000) //8000
+                .setBitrate(8000) //8000
                 .setFrameSize(60)
-                .setNumberOfChannels(2) //1
+                .setNumberOfChannels(1) //1
                 .build();
 
         ScreenShareParam screenShareParam = new ScreenShareParam.Builder()
@@ -265,7 +272,17 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
 
         checkCallPermissions();
 
-        chat.setupCall(videoCallParam, audioCallParam, screenShareParam, callConfig, remoteViews);
+        chat.setupCall(videoCallParam, audioCallParam, screenShareParam, callConfig);
+
+        cpvManager = chat.useCallPartnerViewManager();
+
+        cpvManager.addView(remoteViews);
+
+        cpvManager.setAsCameraPreview(cameraPreview);
+
+        cpvManager.setAsScreenShareView(remoteViews.get(remoteViews.size() - 1));
+
+        cpvManager.setAutoGenerateCallback(this);
 
     }
 
@@ -509,6 +526,11 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     }
 
     @Override
+    public void prepareNewView(CallPartnerView partnerView) {
+        view.addNewView(partnerView);
+    }
+
+    @Override
     public void onStart() {
 
     }
@@ -711,10 +733,10 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     public void onNewMessage(String content, ChatResponse<ResultNewMessage> chatResponse) {
 
 
-        if(CallRecordTracer.isRecordedFileMessage(chatResponse.getUniqueId())){
+        if (CallRecordTracer.isRecordedFileMessage(chatResponse.getUniqueId())) {
 
             try {
-                view.showMessage("تماس در مسیر زیر ذخیره شد: \n "+ chatResponse.getResult().getMessageVO().getMetadata());
+                view.showMessage("تماس در مسیر زیر ذخیره شد: \n " + chatResponse.getResult().getMessageVO().getMetadata());
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -924,9 +946,14 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
                 for (CallParticipantVO cp :
                         callParticipantVOArrayList) {
                     if (cp.hasVideo()) {
-                        CallPartnerView pw = findParticipantView(cp.getUserId());
-                        pw.setPartnerName(cp.getParticipantVO().getContactName());
-                        pw.setDisplayName(true);
+                        if (cpvManager != null) {
+                            cpvManager.showPartnerName(cp.getUserId(), cp.getParticipantVO().getContactName());
+                        } else {
+                            CallPartnerView pw = findParticipantView(cp.getUserId());
+                            pw.setPartnerName(cp.getParticipantVO().getContactName());
+                            pw.setDisplayName(true);
+                        }
+
                     }
                 }
             });
@@ -1240,7 +1267,7 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
 
     }
 
-
+    @SuppressLint("MissingPermission")
     @Override
     public void requestP2PCallWithUserId(int userId, int callType) {
         List<Invitee> invitees = new ArrayList<>();
@@ -1266,6 +1293,7 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
         callImpUniqueIds.add(uniqueId);
     }
 
+    @SuppressLint("MissingPermission")
     @Override
     public void requestP2PCallWithContactId(int contactId, int callType) {
         List<Invitee> invitees = new ArrayList<>();
@@ -1307,7 +1335,7 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
         } else {
             prepareAudioCallView();
         }
-        String uniqueId = chat.requestGroupCall(callRequest);
+        @SuppressLint("MissingPermission") String uniqueId = chat.requestGroupCall(callRequest);
         callUniqueIds.add(uniqueId);
         callImpUniqueIds.add(uniqueId);
 
@@ -1326,11 +1354,12 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
             showVideoViews();
         }
 
-        String uniqueId = chat.requestCall(callRequest);
+        @SuppressLint("MissingPermission") String uniqueId = chat.requestCall(callRequest);
         callUniqueIds.add(uniqueId);
         callImpUniqueIds.add(uniqueId);
     }
 
+    @SuppressLint("MissingPermission")
     @Override
     public void requestMainOrSandboxCall(String query, boolean isGroupCall) {
 
@@ -1435,26 +1464,35 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     @Override
     public void onCallParticipantLeft(ChatResponse<LeaveCallResult> response) {
 
-
         for (CallParticipantVO c :
                 response.getResult().getCallParticipants()) {
             view.showMessage(c.getParticipantVO().getName() + " تماس رو ترک کرد");
         }
-        try {
-            CallPartnerView pw = findParticipantView(response.getResult().getCallParticipants().get(0).getUserId());
-            if (pw != null) {
-                pw.setId(0);
-                pw.setPartnerName("");
-                pw.setDisplayName(false);
-                pw.setDisplayIsMuteIcon(false);
-                pw.setDisplayCameraIsOffIcon(false);
-                pw.reset();
-                chat.addPartnerView(pw, 0);
-            }
-        } catch (Exception e) {
-            view.onError(e.getMessage());
-        }
+        activity.runOnUiThread(() -> {
+            try {
+                if (cpvManager != null) {
+                    cpvManager.releasePartnerView(
+                            response.getResult()
+                                    .getCallParticipants().get(0).getUserId());
 
+                } else {
+                    CallPartnerView pw = findParticipantView(response.getResult().getCallParticipants().get(0).getUserId());
+                    if (pw != null) {
+                        pw.setId(0);
+                        pw.setPartnerName("");
+                        pw.setDisplayName(false);
+                        pw.setDisplayIsMuteIcon(false);
+                        pw.setDisplayCameraIsOffIcon(false);
+                        pw.reset();
+                        chat.addPartnerView(pw, 0);
+                    }
+                }
+
+            } catch (Exception e) {
+                view.onError(e.getMessage());
+            }
+
+        });
 
     }
 
@@ -1465,6 +1503,10 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
 
             try {
                 activity.runOnUiThread(() -> {
+                    if (cpvManager != null) {
+                        cpvManager.showPartnerName(callParticipant.getUserId(), callParticipant.getParticipantVO().getName());
+                        return;
+                    }
                     CallPartnerView pw = findParticipantView(callParticipant.getUserId());
                     pw.setDisplayName(true);
                     pw.setPartnerName(callParticipant.getParticipantVO().getName());
@@ -1472,7 +1514,6 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
             } catch (Exception exc) {
                 Log.e(TAG, "Err: " + exc.getMessage());
             }
-
             view.showMessage(callParticipant.getParticipantVO().getName() + " وارد تماس شد!");
         }
     }
@@ -1526,13 +1567,17 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
                 response.getResult().getCallParticipants()) {
 
             Long userId = participant.getUserId();
-            CallPartnerView partnerView = null;
-            if (userId != null) {
-                partnerView = findParticipantView(userId);
+            if (cpvManager != null) {
+                activity.runOnUiThread(() -> cpvManager.hideMuteIcon(userId));
+            } else {
+                CallPartnerView partnerView = null;
+                if (userId != null) {
+                    partnerView = findParticipantView(userId);
+                }
+                String name = Util.isNotNullOrEmpty(participant.getParticipantVO().getContactName()) ? participant.getParticipantVO().getContactName() : participant.getParticipantVO().getName();
+                view.showMessage(name + " میکروفنش رو باز کرد 😍");
+                view.callParticipantUnMuted(participant, partnerView);
             }
-            String name = Util.isNotNullOrEmpty(participant.getParticipantVO().getContactName()) ? participant.getParticipantVO().getContactName() : participant.getParticipantVO().getName();
-            view.showMessage(name + " میکروفنش رو باز کرد 😍");
-            view.callParticipantUnMuted(participant, partnerView);
         }
     }
 
@@ -1556,13 +1601,17 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
         for (CallParticipantVO participant :
                 response.getResult().getCallParticipants()) {
             Long userId = participant.getUserId();
-            CallPartnerView partnerView = null;
-            if (userId != null) {
-                partnerView = findParticipantView(userId);
+            if (cpvManager != null) {
+                activity.runOnUiThread(() -> cpvManager.showMuteIcon(userId));
+            } else {
+                CallPartnerView partnerView = null;
+                if (userId != null) {
+                    partnerView = findParticipantView(userId);
+                }
+                String name = Util.isNotNullOrEmpty(participant.getParticipantVO().getContactName()) ? participant.getParticipantVO().getContactName() : participant.getParticipantVO().getName();
+                view.showMessage(name + " میکروفنش رو بست 🤐");
+                view.callParticipantMuted(participant, partnerView);
             }
-            String name = Util.isNotNullOrEmpty(participant.getParticipantVO().getContactName()) ? participant.getParticipantVO().getContactName() : participant.getParticipantVO().getName();
-            view.showMessage(name + " میکروفنش رو بست 🤐");
-            view.callParticipantMuted(participant, partnerView);
         }
     }
 
@@ -1580,6 +1629,11 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     public void onCallParticipantStoppedVideo(ChatResponse<JoinCallParticipantResult> response) {
         try {
             view.showMessage(response.getResult().getJoinedParticipants().get(0).getParticipantVO().getName() + " دوربینش رو بست! ");
+            if (cpvManager != null) {
+                activity.runOnUiThread(() -> cpvManager.releasePartnerView(
+                        response.getResult().getJoinedParticipants().get(0).getUserId()));
+                return;
+            }
             CallPartnerView pw = findParticipantView(response.getResult().getJoinedParticipants().get(0).getUserId());
             activity.runOnUiThread(() -> {
                 if (pw != null) {
@@ -1638,7 +1692,11 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
 
     @Override
     public void onCallParticipantStoppedScreenSharing(ChatResponse<ScreenShareResult> response) {
-        view.onCallParticipantStoppedScreenSharing();
+        if (cpvManager != null) {
+            MainThreadExecutor exec = new MainThreadExecutor();
+            exec.execute(() -> cpvManager.releaseScreenShareView());
+        } else
+            view.onCallParticipantStoppedScreenSharing();
     }
 
     @Override
@@ -1680,7 +1738,7 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     @Override
     public void onCallParticipantRecordStopped(ChatResponse<Participant> response) {
         CallRecordTracer.ongoingRecordUniqueId = response.getUniqueId();
-        view.showMessage(" " + response.getResult().getName() + " ضبط تماس را متوقف کرد." );
+        view.showMessage(" " + response.getResult().getName() + " ضبط تماس را متوقف کرد.");
 
     }
 
@@ -1688,7 +1746,7 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
     public void onCallClientErrors(ChatResponse<CallClientErrorsResult> response) {
         CallErrorVO error = response.getResult().getCallErrorVO();
         Participant cp = error.getParticipant();
-        if(error.getCode()== ChatConstant.ERROR_CODE_CAMERA_NOT_AVAILABLE){
+        if (error.getCode() == ChatConstant.ERROR_CODE_CAMERA_NOT_AVAILABLE) {
             view.showMessage("دوربین " + cp.getName() + " قابل استفاده نیست!");
         }
         if (error.getCode() == ChatConstant.ERROR_CODE_MICROPHONE_NOT_AVAILABLE) {
@@ -1726,5 +1784,10 @@ public class CallPresenter extends ChatAdapter implements CallContract.presenter
 //        }
 //        presenter.rejectIncomingCall();
 //        onCallEnded();
+    }
+
+    @Override
+    public void onNewViewGenerated(CallPartnerView callPartnerView) {
+        view.addNewView(callPartnerView);
     }
 }